pyproject.toml
[project]
name = "mcp-server-google-ads"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-auth>=2.41.1",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "mcp>=1.16.0",
    "python-dotenv>=1.1.1",
]
